Bulk PUT is not efficient
Problem
-------

The bulk PUT operation unregisters all of the existing data before
writing the new data.  It is envisioned that very few of the instances
will change so it would be more efficient to only write those that are
new or have changed and only unregister those that need to be
unregistered.

Solution
--------

Compare the new samples to the existing samples and perform the
necessary operation.

Note
----

This can't be merged until OpenDDS generates `==` for the C++11
mapping.