HSDS writer and reader requests are unauthenticated
Problem
-------

The API's exposed by the HSDS do not authenticate callers.

Solution
--------

Require a JWT and ensure that the caller is same as the DPM
application of the writer/reader.

Technically, the API verifies the JWT using the certificate from the
DPM.  It then checks that the principal is the same as for the
credentials used to download DDS security documents from the DPM.

Note
----

TLS is not supported (yet).  Users should still run the writer and
reader with a reverse proxy that can provide TLS termination.